Changed readScout to have a raw result, stats, and comments.
readScout now has 3 variables set at the end that are returned: Result, which is a raw dataframe of the raw data from the json; stats, which is the Result with the matchNo, teamNo, scoutTeamNo, and comments dropped so that it can find the mean, percentiles, min, and max of it via describe( ); and comments, which keeps the teamNo and the comments so one can look at only comments.
Furthermore, a block of code to print out Result, stats, and comments (printed out fully) has been commented out, as well as the option to include scoutName in the comments dataFrame.