in auto picklist graphs, changed boxplot() to plt.boxplot so it doesn't error
In the main function, option 10 lead you to picklist plots. Option 0 fromo that gets you to auto graph sheet, and in the code there was use of boxplot(). However, this causes errors. as boxplot() is undefined. What probably was supposed to happen was that boxplot() is from matplotlib.pyplot, so I changed boxplot() to plt.boxplot(). I also checked that boxplot() is, indeed, a function in the matplotlib.pyplot module.